:recycle: type 변경
ㅇㅇ

+++ src/main/java/com/g3c1/temi/domain/seat/presentation/dto/response/SeatInfoResponse.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class SeatInfoResponse {
     private final Long seatId;
-    private final int seatNumber;
-    private final int severalPeople;
+    private final Long seatNumber;
+    private final Long severalPeople;
     private final Boolean enabled;
 }
